Fixes Issue #15 by converting the topic partials to HAML
Pretty much a straight conversion from ERB to HAML, although I made a
few corrections, like linking the topic name in the list view to the
show page along the way.

I didn't correct some existing deficiencies like form errors not being
displayed or search status not being populated into the page since that
seemed to be out of the scope of the issue.