Remove list_members_path and get member sidebar working
It turns out that Rails 3 is pickier about matching content_for to
yield and wants them to both either be a symbol or a string.

And the list collection action was no longer present in the members
controller but was still in a sidebar and the routes filed, so I
removed it.